stockdox/requirements.txt
gunicorn==20.0.4
django-heroku==0.3.1
whitenoise
asgiref==3.2.3
blessed==1.17.0
botocore==1.13.50
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
Django==3.0
django-crispy-forms==1.8.1
django-maskurl==0.1
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
idna==2.7
jmespath==0.9.4
jsonschema==2.6.0
mysqlclient==1.4.6
pathspec==0.5.9
python-dateutil==2.8.0
pytz==2019.3
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
sqlparse==0.3.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
stripe==2.43.0
six==1.14.0